requirements.txt
tensorflow==2.13.0
keras==2.13.1
keras-vggface==0.6
mtcnn==0.1.1
numpy==1.23.5
opencv-python==4.8.0.76
pillow==10.0.0
scikit-learn==1.3.0
streamlit==1.27.2


